+++ public/src/scripts/exit.js
